+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- added receiver, complement on collect step
+- remove styles (there are in tfg-custom-checkout)
 
+## [0.1.2] - 2022-XX-XX
 ### Added
 - added receiver, complement and suburb with validation and save
 - added google autocomplete view